+++ bindings/js/src/loader.ts
@@ -118,6 +118,7 @@ const metadataSymbols = {
 const describeSymbols = {
   quiver_database_describe_collection: { args: [P, BUF, P], returns: I32 },
   quiver_database_summarize_collection: { args: [P, BUF, P], returns: I32 },
+  quiver_database_number_of_elements: { args: [P, BUF, P], returns: I32 },
 } as const;
 
 const timeSeriesSymbols = {